moved all related components for StudentView folder only
